Some structural refactor
