+++ src/App.css
@@ -53,6 +53,21 @@ body {
   text-align: center;
 }
 
+.button {
+  color: #fff;
+  text-shadow: 0 -1px 0 rgba(0, 0, 0, 0.12);
+  -webkit-box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.045);
+  box-shadow: 2px 2px 2px rgba(0, 0, 0, 0.045);
+  touch-action: manipulation;
+  padding: 4px 15px;
+  font-size: 14px;
+  border-radius: 2px;
+  text-align: center;
+  cursor: pointer;
+  font-weight: 400;
+  line-height: 1.5715;
+}
+
 :root {
   --success-color: #00ae4e;
   --warning-color: #f1c40f;
