--- src/components/table/style.css
@@ -1,13 +1,0 @@
-.table {
-  margin: 0 auto;
-  table-layout: auto;
-  border-collapse: collapse;
-}
-
-.table th,
-.table td {
-  padding: 15px;
-  text-align: left;
-  min-width: 120px;
-  border-bottom: 1px solid #ddd;
-}
