+++ src/pages/initiatives/index.tsx
@@ -1,99 +1,14 @@
 import React from "react";
 import moment from "moment";
 import {Row, Col, Typography} from "antd";
-import {DetailModal} from "../../components/modal";
-import {Table, TableRow} from "../../components/table";
-import data from "../../latest.json";
+import {DetailModal} from "src/components/modal/GroupDetailModal";
+import {InitiativeTable} from "src/components/table/InitiativeTable";
+import {INITIATIVE_TABLE_COLUMNS} from "src/components/table/InitiativeTableColumns";
+import data from "src/latest.json";
 import "./style.css";
 import {Link} from "react-router-dom";
 
 const {Title, Paragraph} = Typography;
-
-const columns = [
-  {
-    key: "_id",
-    dataIndex: "_id",
-    title: "Mongo ID",
-    display: false,
-  },
-  {
-    key: "id",
-    dataIndex: "id",
-    title: "ID",
-    display: false,
-  },
-  {
-    key: "scrapeID",
-    dataIndex: "scrapeID",
-    title: "Scrape Batch ID",
-    display: false,
-  },
-  {
-    key: "modal",
-    dataIndex: "modal",
-    title: "",
-  },
-  {
-    key: "name",
-    dataIndex: "name",
-    title: "",
-  },
-  {
-    key: "description",
-    dataIndex: "description",
-    title: "Description",
-    responsive: ["sm"],
-  },
-  {
-    key: "locations",
-    dataIndex: "locations",
-    title: "Locations",
-    responsive: ["sm"],
-  },
-  {
-    key: "scrapedAt",
-    dataIndex: "scrapedAt",
-    title: "Last Updated",
-    responsive: ["lg"],
-  },
-  {
-    key: "foundedOn",
-    dataIndex: "foundedOn",
-    title: "Created",
-    responsive: ["lg"],
-  },
-  {
-    key: "memberCount",
-    dataIndex: "memberCount",
-    title: "Members",
-    responsive: ["xl"],
-    sorter: (a: TableRow, b: TableRow) => a.memberCount - b.memberCount,
-  },
-  {
-    key: "memberCountIncreaseWeekly",
-    dataIndex: "memberCountIncreaseWeekly",
-    title: "Weekly Growth",
-    responsive: ["xxl"],
-    sorter: (a: TableRow, b: TableRow) =>
-      a.memberCountIncreaseWeekly - b.memberCountIncreaseWeekly,
-  },
-  {
-    key: "postCountIncreaseDaily",
-    dataIndex: "postCountIncreaseDaily",
-    title: "Daily Posts",
-    responsive: ["xl"],
-    sorter: (a: TableRow, b: TableRow) =>
-      a.postCountIncreaseDaily - b.postCountIncreaseDaily,
-  },
-  {
-    key: "postCountIncreasMonthly",
-    dataIndex: "postCountIncreaseMonthly",
-    title: "Monthly Posts",
-    responsive: ["xxl"],
-    sorter: (a: TableRow, b: TableRow) =>
-      a.postCountIncreaseMonthly - b.postCountIncreaseMonthly,
-  },
-];
 
 const filter = (input: string) => {
   const MIN_LENGTH = 30;
@@ -129,8 +44,8 @@ export const InitiativeDirectory = () => (
     </Row>
     <Row gutter={{xs: 8, sm: 16, md: 24, lg: 32}}>
       <Col>
-        <Table
-          columns={columns.filter((col) => col.display !== false)}
+        <InitiativeTable
+          columns={INITIATIVE_TABLE_COLUMNS.filter((col) => col.display !== false)}
           data={data.map((entry) => ({
             ...entry,
             modal: (
